+++ pretty_errors/__init__.py
@@ -1,4 +1,4 @@
-import sys, re, colorama, os, time, linecache, site
+import sys, re, colorama, os, time, linecache
 colorama.init()
 
 name = "pretty_errors"
@@ -703,243 +703,6 @@ def activate():
 
 if _active:
     activate()
-
-
-
-def install(find = False, add_to_user = False, add_to_site = False, pth = False, path = None, clean = None):
-    """Install pretty_errors so that it is imported whenever you run a python file."""
-
-    check = re.compile(r'^\s*import\s+\bpretty_errors\b', re.MULTILINE)
-
-    def readfile(path):
-        try:
-            return ''.join((x for x in open(path)))
-        except IOError:
-            return ''
-
-    def find_install(quiet = False):
-        found = False
-        for path in getallsitepackages():
-            for filename in 'usercustomize.py', 'sitecustomize.py', 'pretty_errors.pth':
-                filepath = os.path.join(path, filename)
-                if check.search(readfile(filepath)):
-                    if not found:
-                        print('\npretty_errors found in:')
-                        found = True
-                    print(filepath)
-        if not found and not quiet:
-            print('\npretty_errors not currently installed in any expected locations.')
-        return found
-
-    def clean_install():
-        pys = []
-        pths = []
-        for path in getallsitepackages():
-            for filename in 'usercustomize.py', 'sitecustomize.py', 'pretty_errors.pth':
-                filepath = os.path.join(path, filename)
-                if check.search(readfile(filepath)):
-                    if filename.endswith('.pth'):
-                        pths.append(filepath)
-                    else:
-                        pys.append(filepath)
-        if not pys and not pths:
-            print('\npretty_errors not currently installed in any expected locations.')
-            return
-        if pths:
-            print('\nAttempting to remove the following files:\n')
-            for filepath in pths:
-                print(' ' + filepath)
-                try:
-                    os.remove(filepath)
-                except Exception as e:
-                    print(e)
-        if pys:
-            print('\nFound entry for pretty_errors in the following files.  ' +
-                'Please edit and remove the relevant section:\n')
-            for filepath in pys:
-                print(' ' + filepath)
-
-    _in_virtualenv = 'VIRTUAL_ENV' in os.environ
-    # in virtualenv, `site` has no attribute `getsitepackages` or `getusersitepackages`
-
-    if _in_virtualenv and add_to_user:
-        print('You are now in virtualenv, which has sitepackage but no user sitepackages, '
-                'thus argv `-u` is incompatible with virtualenv.')
-        print('Please use `-s` instead')
-        return
-
-    def venv_sitepackages():
-        pattern = r'^%s$' % os.path.join(os.environ['VIRTUAL_ENV'], 'lib', 'python[0-9.]+', 'site-packages')
-        sitepackages = [path for path in set(sys.path) if re.search(pattern, path)]
-        return sitepackages
-
-    def getsitepackages():
-        return venv_sitepackages() if _in_virtualenv else site.getsitepackages()
-
-    def getusersitepackages():
-        return [] if _in_virtualenv else [site.getusersitepackages()]
-
-    def getallsitepackages():
-        print('getsitepackages()')
-        print(getsitepackages())
-        print('getusersitepackages()')
-        print(getusersitepackages())
-        return getsitepackages() + getusersitepackages()
-
-
-    if find:
-        find_install()
-        return
-    elif clean:
-        clean_install()
-        return
-
-    if add_to_user:
-        path = getusersitepackages()[0]
-        filename = 'pretty_errors.pth' if pth else 'usercustomize.py'
-    elif add_to_site:
-        path = getsitepackages()[0]
-        filename = 'pretty_errors.pth' if pth else 'sitecustomize.py'
-    else:
-
-        found = find_install(True)
-
-        def get_choice(query, choices, default = None):
-            options = {}
-            for i in range(len(choices)):
-                options[str(i + 1)] = i
-            while True:
-                print()
-                print(' ' + query)
-                print()
-                if found:
-                    print('\nC: Clean startup files (do so before uninstalling pretty_errors)')
-                for option, choice in enumerate(choices):
-                    print('%d: %s' % ((option + 1), choice))
-                print('0: Exit')
-                if default is None:
-                    print('\nOption: ', end='')
-                else:
-                    print('\nOption: [default: %d] ' % (default + 1), end='')
-                choice = input()
-                if choice == '' and default is not None:
-                    choice = str(default + 1)
-                if choice == '0':
-                    sys.exit(0)
-                elif choice.lower() == 'c':
-                    clean_install()
-                    sys.exit(0)
-                elif choice in options:
-                    return options[choice]
-
-        print("""
-To have pretty_errors be used when you run any python file you may add it to your \
-usercustomize.py (user level) or sitecustomize.py (system level), or to pretty_errors.pth.
-
-(just hit <enter> to accept the defaults if you are unsure)
- """)
-
-        paths = getallsitepackages()
-        path = paths[get_choice('Choose folder to install into:', paths, -1 if found else len(paths) - 1)]
-
-        if path in getsitepackages():
-            filenames = ['sitecustomize.py', 'pretty_errors.pth']
-        else:
-            filenames = ['usercustomize.py', 'pretty_errors.pth']
-        filename = filenames[get_choice('Choose file to install into:', filenames, 0)]
-
-    if filename.endswith('.pth'):
-        output = (os.path.dirname(os.path.dirname(os.path.normpath(__file__))) +
-            '\nimport pretty_errors; ' +
-            '#pretty_errors.configure()  ' +
-            '# keep on one line, for options see ' +
-            'https://github.com/onelivesleft/PrettyErrors/blob/master/README.md'
-        )
-    else:
-        output = []
-        output.append('''
-
-###########################################################################
-
-# pretty-errors package to make exception reports legible.
-try:
-    import pretty_errors
-except ImportError:
-    print(
-        'You have uninstalled pretty_errors but it is still present in your python startup.' +
-        '  Please remove its section from file:\\n ' + __file__ + '\\n'
-    )
-else:
-    pass
-
-    # Use if you do not have a color terminal:
-    #pretty_errors.mono()
-
-    # Use to hide frames whose file begins with these paths:
-    #pretty_errors.blacklist('/path/to/blacklist', '/other/path/to/blacklist', ...)
-
-    # Use to only show frames whose file begins with these paths:
-    #pretty_errors.whitelist('/path/to/whitelist', '/other/path/to/whitelist', ...)
-
-    # Use to selectively set a config based on the path to the code of the current frame.
-    #alternate_config = pretty_errors.config.copy()
-    #pretty_errors.pathed_config(alternate_config, '/use/alternate/for/this/path')
-
-    # Use to configure output:
-"""pretty_errors.configure(
-    ''')
-
-        options = []
-        colors = []
-        parameters = []
-        max_length = 0
-        for option in dir(config):
-            if len(option) > max_length:
-                max_length = len(option)
-            if (option not in ('configure', 'mono', 'copy') and
-                    not option.startswith('_')):
-                if option.endswith('_color'):
-                    colors.append(option)
-                else:
-                    options.append(option)
-        indent = '        '
-        for option in sorted(options):
-            if option == 'filename_display':
-                parameters.append(indent + option.ljust(max_length) + ' = pretty_errors.FILENAME_COMPACT,  # FILENAME_EXTENDED | FILENAME_FULL')
-            elif option == 'timestamp_function':
-                parameters.append(indent + option.ljust(max_length) + ' = time.perf_counter')
-            else:
-                parameters.append(indent + option.ljust(max_length) + ' = ' + repr(getattr(config, option)))
-        for option in sorted(colors):
-            parameters.append(indent + option.ljust(max_length) + ' = ' + repr(getattr(config, option)))
-
-        output.append(',\n'.join(parameters))
-        output.append('    )"""\n')
-        output.append('###########################################################################\n')
-        output = '\n'.join(output)
-
-    print('\n--------------')
-
-    filepath = os.path.join(path, filename)
-    if check.search(readfile(filepath)):
-        print('\npretty_errors already present in:\n\n ' + filepath +
-            '\n\nEdit it to set config options.\n')
-        return
-
-    try:
-        os.makedirs(path)
-    except Exception:
-        pass
-    try:
-        out = open(filepath, 'a')
-        out.write(output)
-        out.close()
-    except Exception:
-        print('\nFailed to write to:\n' + filepath)
-    else:
-        print('\npretty_errors added to:\n\n %s\n\nEdit it to set config options.\n' % filepath)
-        if filepath.endswith('.pth'):
-            print('\n*** Delete this file when you uninstall pretty_errors! ***\n')
